app/requirements.txt
prometheus_client
playwright
openpyxl
python-logging-loki